Add support for Watch and TV
Motivation:

This product should be available on as many platforms as possible.

Modifications:

Add minimum OS versions.
Extend to Int64 when necessary to work around smaller Int on watch.

Result:

Product builds for Watch and TV.

+++ Sources/ObliviousHTTP/Varint.swift
@@ -51,7 +51,7 @@ extension ByteBuffer {
 
     @discardableResult
     mutating func writeVarint(_ value: Int) -> Int {
-        switch value {
+        switch Int64(value) {
         case 0..<63:
             // Easy, store the value. The top two bits are 0 so we don't need to do any masking.
             return self.writeInteger(UInt8(truncatingIfNeeded: value))
